changed ssn to unsigned long

+++ include/definitions.h
@@ -22,14 +22,14 @@ enum ERROR_CODES
 
 
 typedef struct {
-  int ssn;                      // Social Security Number
+  unsigned long ssn;             // Social Security Number
   char name[MAX_PATIENT_NAME];   // Patient Name
-  int arrivalTime;             // Arrival Time
+  int arrivalTime;              // Arrival Time
   long arrivalDate;             // Arrival Date
-  int departureTime;           // Departure Time
-  long departureDate;          // Departure Date
+  int departureTime;            // Departure Time
+  long departureDate;           // Departure Date
   char infectious;              // Infectious[Y/N]
-  int seatingNumber;           // Seating Number
+  int seatingNumber;            // Seating Number
 } PatientRecord;
 
 typedef struct{
